Cythonize JunctionCountAssembler and remove old files
