+++ include/khmer/_cpy_khmer.hh
@@ -85,8 +85,6 @@ Contact: khmer-project@idyll.org
 #include "_cpy_nodegraph.hh"
 #include "_cpy_smallcountgraph.hh"
 
-#include "_cpy_assembly.hh"
-
 #include "_cpy_graphlabels.hh" 
 #include "_cpy_hashset.hh" 
 #include "_cpy_hllcounter.hh"
